requirements.txt
filetype
hypercorn[trio]
fastapi
mutagen
pywebview
aiofiles
typer